This commit adds a method called update_post_comments_counter to the Comment model. This method updates the comments_counter column of the associated post. The method is called after the Comment is created or destroyed.
The commit also adds a method called update_likes_counter to the Like model. This method updates the likes_counter column of the associated post. The method is called after the Like is created or destroyed.

Finally, the commit removes the update_comments_counter and update_likes_counter methods from the Post model.